fix(plugin,server,app): serve plugin UI bundles with server-side Solid runtime import rewrite
Move Solid runtime import rewriting from the app client (blob URL) to the
server asset route. This lets Electron production builds import plugin UI
bundles directly under the existing CSP instead of blocking blob: scripts.

- packages/plugin/src/loader.ts: shared rewrite helpers exported from plugin SDK
- packages/plugin/src/index.ts: re-export loader helpers
- packages/synergy/src/server/plugin-routes.ts: rewrite .js/.mjs/.cjs plugin UI bundles on serve
- packages/app/src/plugin/loaders.ts: prefer server-rewritten URL, fall back to blob for older servers
- packages/app/src/plugin/host.tsx, workbench-surface.tsx: remove temporary debug logging
